+++ test/api/info.spec.ts
@@ -9,12 +9,8 @@ describe('API - info', () => {
   })
 
   it('should be rejected with bad options', (done) => {
-    info(factory.uuid('image'), {publicKey: ''})
+    info(factory.uuid('image'), {publicKey: factory.publicKey('empty')})
       .then(() => done.fail())
-      .catch(error => {
-        (error.name === 'UploadcareError')
-          ? done()
-          : done.fail(error)
-      })
+      .catch((error) => error.name === 'UploadcareError' ? done() : done.fail(error))
   })
 })